Refactored components to separate tracking from markup logic
