+++ src/StaticFAQ/__test__/FAQArticle.test.js
@@ -4,7 +4,8 @@ import * as React from 'react';
 import { MemoryRouter } from 'react-router';
 import { mount } from 'enzyme';
 
-import { RawFAQArticle as FAQArticle } from '../FAQArticle';
+//import { RawFAQArticle as FAQArticle } from '../FAQArticle';
+import TrackedFAQArticle from '../TrackedFAQArticle';
 import { SearchState } from '../../context/SearchState';
 
 const mockRefType: any = null;
@@ -32,7 +33,7 @@ describe('FAQArticle', () => {
             incrementQueriesCount,
           }}
         >
-          <FAQArticle
+          <TrackedFAQArticle
             isSearchResult={false}
             article={article}
             categoryId="444"
