+++ src/StaticFAQ/index.js
@@ -1,127 +1,5 @@
 // @flow
 
-import * as React from 'react';
-import idx from 'idx';
-import css from 'styled-jsx/css';
-import { Search } from '@kiwicom/orbit-components/lib/icons';
-import { withRouter } from 'react-router-dom';
+import TrackedStaticFAQ from './TrackedStaticFAQ';
 
-import { SearchState, type SearchStateType } from '../context/SearchState';
-import Input from './../common/Input';
-import FAQCategoryList from './FAQCategoryList';
-import SearchAllFAQs from './SearchAllFAQs';
-import { simpleTracker } from '../helpers/analytics/trackers';
-import { debounce } from '../helpers/functionUtils';
-
-const style = css`
-  .static-faq {
-    width: 100%;
-    height: 100%;
-  }
-  .static-faq-body {
-    height: 100%;
-    padding: 24px 40px;
-    display: flex;
-    flex: 1;
-    flex-direction: column;
-  }
-
-  @media only screen and (max-width: 1180px) {
-    .static-faq-body {
-      padding: 0;
-    }
-    .static-faq-search {
-      margin: 16px;
-    }
-  }
-`;
-
-type Props = {|
-  match: {
-    params: {
-      categoryId: ?string,
-    },
-  },
-|};
-
-type State = {|
-  searchText: string,
-|};
-const logSearchQuery = debounce(q =>
-  simpleTracker('smartFAQBookingOverview', {
-    action: 'search',
-    searchedText: q,
-  }),
-);
-class StaticFAQ extends React.Component<Props, State> {
-  changeSearchText = (text: string) => {}; // eslint-disable-line no-unused-vars
-  resetQueriesCount = () => {}; // eslint-disable-line no-unused-vars
-  incrementQueriesCount = () => {}; // eslint-disable-line no-unused-vars
-
-  handleSearchChange = (e: SyntheticInputEvent<HTMLInputElement>) => {
-    this.changeSearchText(e.target.value);
-    logSearchQuery(e.target.value);
-  };
-  handleCancelSearch = () => {
-    this.changeSearchText('');
-    this.resetQueriesCount();
-  };
-
-  renderInput = (searchText: string) => {
-    const isSearching = searchText.length > 0;
-    const inputChange = e => {
-      this.incrementQueriesCount();
-      this.handleSearchChange(e);
-    };
-    return (
-      <Input
-        value={searchText}
-        onChange={inputChange}
-        placeholder="What can we help you with?"
-        icon={<Search customColor="#bac7d5" />}
-        onReset={isSearching ? this.handleCancelSearch : undefined}
-        dataCy="input-staticFAQ"
-      />
-    );
-  };
-
-  render() {
-    const categoryId = idx(this.props.match, _ => _.params.categoryId) || null;
-
-    return (
-      <SearchState.Consumer>
-        {({
-          searchText,
-          changeSearchText,
-          incrementQueriesCount,
-          resetQueriesCount,
-        }: SearchStateType) => {
-          this.changeSearchText = changeSearchText;
-          this.resetQueriesCount = resetQueriesCount;
-          this.incrementQueriesCount = incrementQueriesCount;
-          const isSearching = searchText.length > 0;
-
-          return (
-            <div className="static-faq">
-              <div className="static-faq-body">
-                {!categoryId && (
-                  <div className="static-faq-search">
-                    {this.renderInput(searchText)}
-                  </div>
-                )}
-                {isSearching ? (
-                  <SearchAllFAQs search={searchText} />
-                ) : (
-                  <FAQCategoryList categoryId={categoryId} />
-                )}
-              </div>
-              <style jsx>{style}</style>
-            </div>
-          );
-        }}
-      </SearchState.Consumer>
-    );
-  }
-}
-
-export default withRouter(StaticFAQ);
+export default TrackedStaticFAQ;
